WEBAPP-393: Remove redux dependencies from endpoint package
