+++ src/modules/app/actions/__tests__/finishFetchAction.spec.js
