+++ src/modules/app/createReduxStore.js
@@ -8,7 +8,7 @@ import { connectRoutes } from 'redux-first-router'
 import { createLogger } from 'redux-logger'
 
 import uiDirectionReducer from '../i18n/reducers'
-import endpointReducers from '../endpoint/reducers'
+import endpointReducers from './reducers'
 import toggleDarkModeReducer from '../theme/reducers'
 import { createResponsiveStateReducer, responsiveStoreEnhancer } from 'redux-responsive'
 import defaultRoutesMap from './routesMap'
