+++ src/modules/app/reducers/index.js
@@ -1,19 +1,19 @@
 // @flow
 
-import languagesEndpoint from '../endpoints/languages'
-import citiesEndpoint from '../endpoints/cities'
-import categoriesEndpoint from '../endpoints/categories'
-import eventsEndpoint from '../endpoints/events'
-import disclaimerEndpoint from '../endpoints/disclaimer'
-import extrasEndpoint from '../endpoints/extras'
-import sprungbrettJobEndpoint from '../endpoints/sprungbrettJobs'
-import wohnenEndpoint from '../endpoints/wohnen'
-import poisEndpoint from '../endpoints/pois'
+import languagesEndpoint from '../../endpoint/endpoints/languages'
+import citiesEndpoint from '../../endpoint/endpoints/cities'
+import categoriesEndpoint from '../../endpoint/endpoints/categories'
+import eventsEndpoint from '../../endpoint/endpoints/events'
+import disclaimerEndpoint from '../../endpoint/endpoints/disclaimer'
+import extrasEndpoint from '../../endpoint/endpoints/extras'
+import sprungbrettJobEndpoint from '../../endpoint/endpoints/sprungbrettJobs'
+import wohnenEndpoint from '../../endpoint/endpoints/wohnen'
+import poisEndpoint from '../../endpoint/endpoints/pois'
 
 import { handleActions } from 'redux-actions'
-import Payload from '../Payload'
-import { startFetchActionName } from '../actions/startFetchAction'
-import { finishFetchActionName } from '../actions/finishFetchAction'
+import Payload from '../../endpoint/Payload'
+import { startFetchActionName } from '../../app/actions/startFetchAction'
+import { finishFetchActionName } from '../../app/actions/finishFetchAction'
 import type { ActionType } from '../../app/createReduxStore'
 
 /**
